chore: make genre carousel items better in light mode
